restructure testing dir #36
