+++ __tests__/App.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import App from "./App";
+import App from "../App";
 
 import renderer from "react-test-renderer";
 
@@ -8,7 +8,7 @@ jest.mock("expo", () => ({
   AppLoading: "AppLoading"
 }));
 
-jest.mock("./navigation/navigation", () => "navigation");
+jest.mock("../navigation/navigation", () => "navigation");
 
 // Snapshot test
 test("renders correctly", () => {
