+++ __tests__/sliding-panel.test.js
@@ -1,7 +1,9 @@
 import React from "react";
-import SlidingPanel from "./sliding-panel.component";
+import SlidingPanel from "../components/sliding-panel/sliding-panel.component";
 import renderer from "react-test-renderer";
 
+// Kind: Snapshot + Unit Test
+// [US-4] Checks if Sliding Panel renders correctly with the right assets
 describe("Sliding Panel component", () => {
   test("renders correctly", () => {
     const tree = renderer.create(<SlidingPanel />).toJSON();
