--- jimmy-test/sum.js
@@ -1,5 +1,0 @@
-function sum(a, b) {
-  return a + b;
-}
-
-module.exports = sum;
