--- jimmy-test/sum.test.js
@@ -1,9 +1,0 @@
-const sum = require("./sum");
-
-test("adds 1 + 2 to equal 3", () => {
-  expect(sum(1, 2)).toBe(3);
-});
-
-test("adds 1 + -1 to equal 0", () => {
-  expect(sum(1, -1)).toBe(0);
-});
